Add smoke test and refactor src location
