+++ .github/workflows/smoketest.yml
@@ -1,0 +1,20 @@
+name: Smoke test
+
+on:
+  push:
+  schedule:
+    - cron: 0 0 * * *
+
+jobs:
+  smoke-test:
+    name: Smoke test
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v2
+      - uses: actions/setup-node@v1
+        with:
+          node-version: 12
+      -
+        run: |
+          npm ci
+          npm test
