+++ package.json
@@ -2,9 +2,9 @@
   "name": "bookcovers",
   "version": "0.0.2",
   "description": "A federated search for thumbnail images of book covers",
-  "main": "index.js",
+  "main": "src/index.js",
   "bin": {
-    "bookcovers": "cli.js"
+    "bookcovers": "src/cli.js"
   },
   "dependencies": {
     "commander": "^3.0.2",
@@ -15,9 +15,11 @@
     "request": "^2.88.0",
     "request-promise-native": "^1.0.7"
   },
-  "devDependencies": {},
+  "devDependencies": {
+    "jest": "^26.6.3"
+  },
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "jest"
   },
   "author": "e-e-e",
   "license": "ISC"
